requirements.txt
disnake
rethinkdb
aiofiles
aiohttp_client_cache
python-dotenv
loguru
Google-Images-Search
lxml
cssselect
Pillow